Update TicTacToe.java
I have added code that sets the colour red for player X and the colour blue for player O, and code that resets the colours of the buttons to white when the game ends. In this way, a red tile will appear on the board after player X's move and a blue tile will appear after player O's move. At the end of the game, the colours of the tiles will be reset to white.